requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==3.2.0
cffi==1.17.1
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.110.3
greenlet==3.1.1
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.23
starlette==0.37.2
typing_extensions==4.12.2
